RestApi/pyproject.toml
[tool.poetry]
name = "kill-depression"
version = "0.1.0"
description = ""
authors = ["kYRYL"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
requests = "^2.32.3"
openai = "^1.77.0"
supabase = "^2.15.1"
python-dotenv = "^1.1.0"
python-jose = "^3.4.0"
uvicorn = "^0.34.2"
transformers = "^4.51.3"
pytest = "^8.3.5"

[tool.pytest.ini_options]
pythonpath = "."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
